pyproject.toml
[project]
name = "lexigaze"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.73.1",
    "matplotlib>=3.10.9",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "spacy>=3.8.14",
    "transformers>=5.7.0",
]
